backend/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-js-asset==2.0.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg==3.1.7
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
typing_extensions==4.4.0
whitenoise==6.3.0
